% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EffectSizeConversions.R
\name{d_eta2}
\alias{d_eta2}
\title{Convert a Cohen's d to an \eqn{\eta^2_p}.}
\usage{
d_eta2(d)
}
\arguments{
\item{d2}{The Cohen's d value.}
}
\value{
\eqn{\eta^2_p}
}
\description{
Convert a Cohen's d to an \eqn{\eta^2_p}.
}
