% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispersion.R
\name{disp_ratio}
\alias{disp_ratio}
\title{Dispersion ratio}
\usage{
disp_ratio(model, ...)
}
\arguments{
\item{model}{a fitted regression model object that has a relevant pearson residual to be extracted}

\item{...}{dots}
}
\value{
A dispersion ratio, where 1 is equidispersion (as expected), > 1 is over-dispersion and <1 is under-dispersion
}
\description{
This function calculates the dispersion ratio of fitted models. Aimed at binomial and Poisson family models where
The fixed variance assumption commonly leads to over- dispersion in the real world.
}
\examples{
library(NHSRdatasets)
data(LOS_model)

mod1 <- glm(Death ~ Age * LOS, data=LOS_model, family="binomial")

disp_ratio(mod1)
}
