% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EffectSizeConversions.R
\name{eta2_d}
\alias{eta2_d}
\title{Convert an \eqn{\eta^2_p} to a Cohen's d.}
\usage{
eta2_d(eta2)
}
\arguments{
\item{eta2}{The \eqn{\eta^2_p} value.}
}
\value{
Cohen's d.
}
\description{
Convert an \eqn{\eta^2_p} to a Cohen's d.
}
