% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EffectSizeConversions.R
\name{eta2_f}
\alias{eta2_f}
\title{Convert an \eqn{\eta^2_p} to an \eqn{f}.}
\usage{
eta2_f(eta2)
}
\arguments{
\item{eta2}{The \eqn{\eta^2_p} value.}
}
\value{
\eqn{f}
}
\description{
Convert an \eqn{\eta^2_p} to an \eqn{f}.
}
