% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EffectSizeConversions.R
\name{eta2_f2}
\alias{eta2_f2}
\title{Convert an \eqn{\eta^2_p} to an \eqn{\f^2}.}
\usage{
eta2_f2(eta2)
}
\arguments{
\item{eta2}{The \eqn{\eta^2_p} value.}
}
\value{
The \eqn{f^2}.
}
\description{
Convert an \eqn{\eta^2_p} to an \eqn{\f^2}.
}
