% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EffectSizeConversions.R
\name{eta2_r}
\alias{eta2_r}
\title{Convert an \eqn{\eta^2_p} to a correlation coefficient (\eqn{r}).}
\usage{
eta2_r(eta2)
}
\arguments{
\item{eta2}{The \eqn{\eta^2_p} value.}
}
\value{
\eqn{r}
}
\description{
Convert an \eqn{\eta^2_p} to a correlation coefficient (\eqn{r}).
}
