% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EffectSizeConversions.R
\name{f2_eta2}
\alias{f2_eta2}
\title{Convert an \eqn{\f^2} to an \eqn{\eta^2_p}.}
\usage{
f2_eta2(f2)
}
\arguments{
\item{eta2}{The \eqn{f^2} value.}
}
\value{
The \eqn{\eta^2_p}.
}
\description{
Convert an \eqn{\f^2} to an \eqn{\eta^2_p}.
}
