% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EffectSizeConversions.R
\name{f_eta2}
\alias{f_eta2}
\title{Convert an \eqn{f} to an \eqn{\eta^2_p}.}
\usage{
f_eta2(f)
}
\arguments{
\item{f}{The \eqn{f} value.}
}
\value{
\eqn{\eta^2_p}
}
\description{
Convert an \eqn{f} to an \eqn{\eta^2_p}.
}
