% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GroupCenter.R
\name{groupCenter}
\alias{groupCenter}
\title{Group Centering Function}
\usage{
groupCenter(var, group, centerFn = mean, ...)
}
\arguments{
\item{var}{A vector of values.}

\item{group}{A group identifier variable containing a unique value for each
group.}

\item{centerFn}{A function object used to compute the value to subtract
from each score. It should accept a vector (passed by \code{tapply}) and
return a single value.}
}
\description{
This function subtracts the value returned by \code{centerFn} for each
unique \code{group} from \code{var}, and returns the modified vector.
}
