% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mySummaryLm.R
\name{lmSummary}
\alias{lmSummary}
\title{Based on `summary.lm`, this function prints `lm()` objects with extra info.}
\usage{
lmSummary(object, correlation = FALSE, symbolic.cor = FALSE, ...)
}
\arguments{
\item{object}{The `lm` object to report results from.}

\item{correlation}{Prints partial correlations instead of R^2}
}
\description{
Based on `summary.lm`, this function prints `lm()` objects with extra info.
}
