% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadStuff.R
\name{loadStuff}
\alias{loadStuff}
\title{Checks to see if each package in the list is installed. If yes, loads it.
If no, installs it, then loads it.}
\usage{
loadStuff(pkgList)
}
\arguments{
\item{pkgList}{A character vector of package names to install/load}
}
\description{
Checks to see if each package in the list is installed. If yes, loads it.
If no, installs it, then loads it.
}
