% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogitUtil.R
\name{logitToOdds}
\alias{logitToOdds}
\title{Convert a logit unit to an odds ratio.}
\usage{
logitToOdds(l)
}
\arguments{
\item{l}{The logit.}
}
\value{
The odds ratio.
}
\description{
Convert a logit unit to an odds ratio.
}
