% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogitUtil.R
\name{logitToP}
\alias{logitToP}
\title{Convert a logit unit to a probability.}
\usage{
logitToP(l)
}
\arguments{
\item{l}{The logit.}
}
\value{
The probability.
}
\description{
Convert a logit unit to a probability.
}
