% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePdf.R
\name{makePdf}
\alias{makePdf}
\title{Write to a pdf, but allow printing to the default device first.}
\usage{
makePdf(file, plotFn, printToDefaultDev = T, ...)
}
\arguments{
\item{file}{The filename to be passed to pdf()}

\item{plotFn}{A function that wraps the plot statement. Should not take any
arguments.}

\item{printToDefaultDev}{If TRUE (default), the plot function will first
be called without opening the pdf graphics device. Thus, it will print to 
the screen, or in your markdown output, etc.}
}
\description{
Write to a pdf, but allow printing to the default device first.
}
