% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogitUtil.R
\name{oddsToLogit}
\alias{oddsToLogit}
\title{Convert an odds ratio to a logit.}
\usage{
oddsToLogit(o)
}
\arguments{
\item{o}{The odds ratio.}
}
\value{
The logit.
}
\description{
Convert an odds ratio to a logit.
}
