% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogitUtil.R
\name{oddsToP}
\alias{oddsToP}
\title{Convert an odds ratio to a probability.}
\usage{
oddsToP(o)
}
\arguments{
\item{o}{The odds ratio.}
}
\value{
The probability.
}
\description{
Convert an odds ratio to a probability.
}
