% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispersion.R
\name{phi_func}
\alias{phi_func}
\title{Calculate overdispersion ratio of z-scores}
\usage{
phi_func(n, zscores)
}
\arguments{
\item{n}{Single numeric value for the count of the number of groups (and therefore z-scores)}

\item{zscores}{Vector of z-scores z-scores to be used.  Commonly, this might be 'winsorised' first to remove impact of extreme outliers.}
}
\value{
A numeric phi value
}
\description{
Internal function to perform the transformations for data types.
}
\examples{
phi_func(3, c(1.3,0.75, 1.5))

}
