% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pLvlCenter.R
\name{pLvlCenter}
\alias{pLvlCenter}
\title{Center Variables at the Participant Level}
\usage{
pLvlCenter(df, var, group = "pnum", cntrValFn = mean,
  returnType = c("df", "vec"), ...)
}
\arguments{
\item{df}{Data frame.}

\item{var}{Variable in \code{df} to center.}

\item{group}{Grouping variable. Must be in \code{df}!}

\item{cntrValFn}{Function that returns the value to center around within each
participant. Most often the default, \code{mean}.}

\item{returnType}{Can return either a dataframe ("df"; default) or a vector
("vec").}

\item{...}{Other arguments to pass to \code{ctrlValFn}.}
}
\value{
Either a dataframe with the new variable or a vector of the centered
values, depending on the \code{returnType} argument.
}
\description{
Center Variables at the Participant Level
}
