% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidence_intervals.R
\name{prop_ci}
\alias{prop_ci}
\title{Wilson score binomial confidence interval for proportions}
\usage{
prop_ci(o, n, ci = 0.95)
}
\arguments{
\item{o}{observed or numerator value}

\item{n}{expected or denominator value}

\item{ci}{confidence interval coverage required.  Default is 0.95 for 95\% confidence interval}
}
\value{
A vector o/n, lower confidence interval limit, and upper confidence interval limit
}
\description{
A Wilson score-based CI calculation for a proportion.  Based on guidance by PHE
}
\examples{
prop_ci(50,120)
}
