% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogitUtil.R
\name{pToLogit}
\alias{pToLogit}
\title{Convert a probability unit to a logit.}
\usage{
pToLogit(p)
}
\arguments{
\item{p}{The probability.}
}
\value{
The logit.
}
\description{
Convert a probability unit to a logit.
}
