% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogitUtil.R
\name{pToOdds}
\alias{pToOdds}
\title{Convert a probility to an odds ratio.}
\usage{
pToOdds(p)
}
\arguments{
\item{p}{The probability.}
}
\value{
The odds ratio.
}
\description{
Convert a probility to an odds ratio.
}
