% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EffectSizeConversions.R
\name{r_eta2}
\alias{r_eta2}
\title{Convert a correlation coefficient to a \eqn{\eta^2_p}.}
\usage{
r_eta2(r)
}
\arguments{
\item{r}{The correlation coefficient.}
}
\value{
\eqn{\eta^2_p}
}
\description{
Convert a correlation coefficient to a \eqn{\eta^2_p}.
}
