% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Recycle.R
\name{recycle}
\alias{recycle}
\title{Recycle a vector to the specified length.
 
 @param vec The vector to recycle.
 @param length Either a vector whose length will be matched, or an integer
 @param na.fill If TRUE, then the vector is simply filled with NAs to 
 match `length`
 
 @export recycle}
\usage{
recycle(vec, length, na.fill = FALSE)
}
\description{
Recycle a vector to the specified length.
 
 @param vec The vector to recycle.
 @param length Either a vector whose length will be matched, or an integer
 @param na.fill If TRUE, then the vector is simply filled with NAs to 
 match `length`
 
 @export recycle
}
