% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vioplotCM.R
\name{vioplotCM}
\alias{vioplotCM}
\title{violin plots with control}
\usage{
vioplotCM(x, ..., range = 1.5, h = NULL, ylim = NULL, names = NULL,
  horizontal = FALSE, col = "cornflowerblue", border = "black",
  lty = 1, lwd = 1, rectCol = "black", colMed = "white",
  pchMed = 19, at, add = FALSE, wex = 1, drawRect = TRUE,
  na.fun = na.omit, drawMean = T, mnLty = 2, mnLwd = 1.5,
  mnCol = "black", mnLength = NULL, main = "", mainLine = 1,
  maincex = 1.5, nameCex = 1, xlab = "", xlabLine = 2,
  xlabcex = 1.1, ylab = NULL, ylabLine = 2.5, ylabcex = 1.3,
  labelMeans = F, mnsDigits = 2)
}
\arguments{
\item{x}{A vector of values to vioplot.}

\item{...}{More vectors of values to vioplot.}

\item{range}{A factor to calculate the upper/lower adjacent values.}

\item{h}{The height for the density estimator, if omit as explained in 
sm.density, h will be set to an optimum.}

\item{ylim}{y limits (vector of length 2).}

\item{names}{An optional character vector of labels for each set of values to
be vioplotted. If omitted, they are generated from the variable names.}

\item{horizontal}{Logical, make it horizontal or vertical.}

\item{col}{A recycled vector of colors for the vioplots themselves.}

\item{border, rectCol}{Colors for the things.}

\item{lty, lwd}{Graphical paramters when drawing the lines.}

\item{colMed, pchMed}{Graphical paramters for the symbol representing the 
median.}

\item{at}{The position of each violin. Defaults to 1:n, where n is the number
of vectors being vioplotted.}

\item{add}{If FALSE (default) a new plot is created.}

\item{wex}{Relative expansion of the violin.}

\item{drawRect}{Put on a rectangle on it?}

\item{na.fun}{Defaults to \code{na.omit()}. This is the way that \code{NA}s 
are dealt with.}

\item{drawMean}{Should a line be drawn for each mean?}

\item{mnLty, mnLwd, mnCol}{Graphical paramters for drawing the mean line.}

\item{mnLength}{The length of the mean line in plot units. Defaults to 
something sensible.}

\item{main}{Plot main title.}

\item{mainLine, maincex}{Graphical parameters for the main title.}

\item{nameCex}{Relative size of the category labels.}

\item{xlab}{A title for the x-axis placed underneath the category labels.}

\item{xlabLine, xlabCex}{Graphical parameters for the x-axis label.}

\item{ylab}{An optional label for the y-axis.}

\item{ylabLine, ylabCex}{Graphical paramters for the y-axis label.}

\item{labelMeans}{Logical. If true, the means for each group will be
calculated and added in parentheses to the \code{names} items.}

\item{mnsDigits}{The number of digits to round the means values to.}
}
\description{
Plot violin plots the cool way.
}
